Kill some longs.
In portable code, basically every use of long or %ld/%lu/%lx is a bug.
A legitimate use in Crawl is the bit array: storing it in-memory should be
preferably done a word at a time.

+++ crawl-ref/source/cluautil.h
@@ -134,7 +134,7 @@ class monster;
 struct MonsterWrap
 {
     monster* mons;
-    long      turn;
+    int      turn;
 };
 
 // XXX: These are currently defined outside cluautil.cc.
